Factor out NSStatusItem-creation in own class. Now Re-initialize status items whenever an instance is modified. No need to restart app anymore.
